feat(prosemirror): handle texts size
