+++ packages/oak-prosemirror-editor/lib/core/Editor/utils.js
@@ -1,3 +1,8 @@
+export const SIZES = {
+  text: 16,
+  headings: { h1: 32, h2: 24, h3: 19, h4: 16, h5: 13, h6: 10 },
+};
+
 export const getActiveAttrs = (editorState, type) => {
   const { from, to } = editorState.selection;
   let marks = [];
